Move Activity Tracking Mode to ProfileCard, remove blurbs from GoalsSection, update collapsed summary to show height/weight
